Add files via upload
This was my very first attempt at making a project with html, java script and css during my days at a Japanese university. When trying to play this game, please make sure to put all the files under the same domain. Also, the language used in this game is Japanese, and there will be no English option.